Added new scene called Instrument Prototype.
In the scene, I used Interactable and InstrumentObject scripts on the Cube Instrument to make clicking move it to your 'arm'. There is space in this generic InstrumentObject script to activate the control scheme of whichever instrument we please, and to enable recording as well.
