+++ Assets/OverworldPrototype/Scenes/Recording and Music Test.unity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 203ad2bdce438ee45bfa9b9e2d60a28a
+timeCreated: 1506206656
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
