+++ Assets/OverworldPrototype/Scripts/Character/FirstPersonController.cs
@@ -4,7 +4,8 @@ using UnityEngine;
 
 public class FirstPersonController : MonoBehaviour {
 
-    public float speed = 2.0f;
+    public float speed;
+    public float sprintSpeed;
     public float scrollSpeed = 2.0f;
     CharacterController player;
 
@@ -31,6 +32,14 @@ public class FirstPersonController : MonoBehaviour {
     }
 	
 	void Update () {
+        if(Input.GetKey(KeyCode.LeftShift)|| Input.GetKey(KeyCode.RightShift))
+        {
+            speed = sprintSpeed;
+        }
+        else
+        {
+            speed = 10f;
+        }
         moveForwardBackward = Input.GetAxis("Vertical") * speed;
         moveLeftRight = Input.GetAxis("Horizontal") * speed;
         moveUpDown = Input.GetAxis("Mouse ScrollWheel") * scrollSpeed;
@@ -39,9 +48,9 @@ public class FirstPersonController : MonoBehaviour {
 
         movement = transform.rotation * movement;
         player.Move(movement * Time.deltaTime);
-        
-        
 
+
+        Debug.Log(speed);
 
 
     }
