+++ Assets/OverworldPrototype/Scripts/Objects.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d96167a9b158deb49ba520e4fb1601f8
+folderAsset: yes
+timeCreated: 1506207245
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
