+++ Assets/OverworldPrototype/Scripts/Objects/InstrumentObject.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 85df6aefdd7b2334c9e3fbc26afebdc5
+timeCreated: 1506207273
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
