+++ Assets/OverworldPrototype/Scripts/Objects/Interactable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 22baae40fa3ea154a9596c510f21b597
+timeCreated: 1505243196
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
