+++ Assets/Sounds/ClickSound.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: beaf3d2729c57e74abb0e3cd8105f381
+timeCreated: 1506210313
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
